+++ request/utils.py
@@ -11,6 +11,7 @@ HTTP_STATUS_CODES = (
     (100, _('Continue')),
     (101, _('Switching Protocols')),
     (102, _('Processing (WebDAV)')),
+    (103, _('Early Hints')),
 
     # Success
     (200, _('OK')),
@@ -21,6 +22,8 @@ HTTP_STATUS_CODES = (
     (205, _('Reset Content')),
     (206, _('Partial Content')),
     (207, _('Multi-Status (WebDAV)')),
+    (208, _('Already Reported (WebDAV)')),
+    (226, _('IM Used (HTTP Delta encoding)')),
 
     # Redirection
     (300, _('Multiple Choices')),
@@ -31,6 +34,7 @@ HTTP_STATUS_CODES = (
     (305, _('Use Proxy')),
     (306, _('Switch Proxy')),  # No longer used
     (307, _('Temporary Redirect')),
+    (308, _('Permanent Redirect')),
 
     # Client Error
     (400, _('Bad Request')),
@@ -57,7 +61,11 @@ HTTP_STATUS_CODES = (
     (424, _('Failed Dependency (WebDAV)')),
     (425, _('Unordered Collection')),
     (426, _('Upgrade Required')),
+    (428, _('Precondition Required')),
+    (429, _('Too Many Requests')),
+    (431, _('Request Header Fields Too Large')),
     (449, _('Retry With')),
+    (451, _('Unavailable For Legal Reasons')),
 
     # Server Error
     (500, _('Internal Server Error')),
@@ -70,6 +78,7 @@ HTTP_STATUS_CODES = (
     (507, _('Insufficient Storage (WebDAV)')),
     (509, _('Bandwidth Limit Exceeded')),
     (510, _('Not Extended')),
+    (511, _('Network Authentication Required')),
 )